+++ src/shapenodes/SoMarkerSet.cpp
@@ -1189,12 +1189,12 @@ SoMarkerSet::GLRender(SoGLRenderAction * action)
   // view angles. We'll disable the clipping planes temporarily. Individual
   // markers are still clipped using SoCullElement::cullTest() below.
   // See https://bitbucket.org/Coin3D/coin/pull-requests/52 for a test case.
-  int numPlanes = 0;
+  GLint numPlanes = 0;
   glGetIntegerv(GL_MAX_CLIP_PLANES, &numPlanes);
   SbList<SbBool> planesEnabled;
-  for (int i = 0; i < numPlanes; ++i) {
-    planesEnabled.append(glIsEnabled(GL_CLIP_PLANE0 + (GLuint)i));
-    glDisable(GL_CLIP_PLANE0 + (GLuint)i);
+  for (GLint i = 0; i < numPlanes; ++i) {
+    planesEnabled.append(glIsEnabled(GL_CLIP_PLANE0 + i));
+    glDisable(GL_CLIP_PLANE0 + i);
   }
 
   glMatrixMode(GL_MODELVIEW);
@@ -1257,9 +1257,9 @@ SoMarkerSet::GLRender(SoGLRenderAction * action)
     glBitmap(tmp->width, tmp->height, 0, 0, 0, 0, tmp->data);
   }
 
-  for (int i = 0; i < numPlanes; ++i) {
+  for (GLint i = 0; i < numPlanes; ++i) {
     if (planesEnabled[i]) {
-      glEnable(GL_CLIP_PLANE0 + (GLuint)i);
+      glEnable(GL_CLIP_PLANE0 + i);
     }
   }
